Split java type functions to a separate class
